+++ CHANGELOG.md
@@ -2,5 +2,6 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
+* Initial creation of the provider using v0.0.10 of the Confluent Cloud Terraform Provider
 
 ---